refactor: combined utils functions for working with dates into one file
